+++ src/upgrade.js
@@ -1,7 +1,7 @@
 // This file handles various upgrade procedures
 // when file syntax changes between versions.
-const util = require('./util.js');
-const fs = require('fs');
+// const util = require('./util.js');
+// const fs = require('fs');
 
 module.exports = {
   // Takes inputs:
@@ -10,31 +10,6 @@ module.exports = {
   // files, object containing paths to various json files (settings, etc.)
   // callback, which is callback
   upgrade: function (fromVersion, toVersion, files) {
-    // files is an object taking values:
-    // customColorFile
-
-    // Handle color files format change in 1.1.2 => 1.1.3
-    if (!util.semVerGreaterThan(fromVersion, '1.1.3')) {
-      const colorData = JSON.parse(fs.readFileSync(files.customColorFile));
-      // Prevent modification multiple times
-      if (colorData.colors) {
-        return;
-      }
-
-      fs.unlinkSync(files.customColorFile);
-      // Create a color scheme file with the color data
-      const jsonTemplateC = {
-        customColorScheme: {
-          name: 'Custom color scheme',
-          id: 'customColorScheme',
-          colors: colorData
-        }
-      };
-      fs.writeFile(files.customColorFile, JSON.stringify(jsonTemplateC), 'utf8', function (err) {
-        if (err) {
-          console.error(err);
-        }
-      });
-    }
+    // At the moment, not used for anything
   }
 };